+++ test/twg2/arrays/test/ArrayManagerTests.java
@@ -12,17 +12,28 @@ public class ArrayManagerTests {
 	@Test
 	public void expandTest() {
 		int[] ary = new int[] { 0, 1 };
-		int size = ary.length;
 
-		for(int i = size; i < 20; i++) {
-			if(size + 1 > ary.length) {
+		for(int i = ary.length; i < 20; i++) {
+			if(i + 1 > ary.length) {
 				ary = ArrayManager.expand(ary);
 			}
 			ary[i] = ary[i - 2] + ary[i - 1];
-			size++;
 		}
 
 		Assert.assertArrayEquals(new int[] { 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584, 4181 }, Arrays.copyOf(ary, 20));
+
+		String[] strs = new String[] { "A", "B", "C" };
+		int i = 3;
+		for(int a = 0; a < 3; a++) {
+			for(int b = 0; b < 3; b++) {
+				if(i + 1 > strs.length) {
+					strs = ArrayManager.expand(strs);
+				}
+				strs[i] = strs[a] + strs[b];
+				i++;
+			}
+		}
+		Assert.assertArrayEquals(new String[] { "A", "B", "C", "AA", "AB", "AC" }, Arrays.copyOf(strs, 6));
 	}